refactor FailedApplicationServers counting refactor healthchecks (1to5 files)
